cinemachine/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
asyncpg==0.30.0
google-cloud-storage==2.19.0
google-genai==1.5.0
google-auth==2.37.0
python-jose[cryptography]==3.3.0
websockets==14.1
pydantic==2.10.4
pydantic-settings==2.7.1
httpx==0.28.1
